# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server CRM server.",
        "title": "Golang CRM Swagger Documentation",
        "contact": {
            "name": "API Support",
            "url": "https://t.me/xalmatoff",
            "email": "venom.uzz@mail.ru"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/products": {
            "post": {
                "description": "This API to create a product.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Create a product.",
                "parameters": [
                    {
                        "description": "data body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateProductInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Products"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/products-by-filter": {
            "get": {
                "description": "This API to get products with images by filter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get products with images by filter.",
                "parameters": [
                    {
                        "type": "string",
                        "x-order": "1",
                        "name": "nameUz",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "x-order": "10",
                        "name": "all",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "x-order": "11",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "x-order": "12",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "2",
                        "name": "url",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "x-order": "3",
                        "name": "price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "4",
                        "name": "descriptionUz",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "6",
                        "name": "groupUrl",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "x-order": "7",
                        "name": "enabled",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "name_uz",
                            "price",
                            "created_at"
                        ],
                        "type": "string",
                        "x-order": "8",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "x-order": "9",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductsWithImagesAndPagination"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/products-url/{url}": {
            "get": {
                "description": "This API to get product by url.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product by url.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "url for get Product",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductWithImages"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/products/{id}": {
            "get": {
                "description": "This API to get product with images by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product with images by id.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "id for get Product",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductWithImages"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "This API to update a product.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Update a product.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "id for update Product",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProductInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Products"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "This API to delete a product.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "id for delete Product",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateProductInput": {
            "type": "object",
            "required": [
                "descriptionUz",
                "groupId",
                "nameUz",
                "price"
            ],
            "properties": {
                "url": {
                    "type": "string",
                    "x-order": "1"
                },
                "price": {
                    "type": "number",
                    "x-order": "10"
                },
                "seoDescription": {
                    "type": "string",
                    "x-order": "11"
                },
                "seoKeywords": {
                    "type": "string",
                    "x-order": "12"
                },
                "seoText": {
                    "type": "string",
                    "x-order": "13"
                },
                "seoTitle": {
                    "type": "string",
                    "x-order": "14"
                },
                "enabled": {
                    "type": "boolean",
                    "x-order": "15"
                },
                "nameUz": {
                    "type": "string",
                    "x-order": "2"
                },
                "nameRu": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameEn": {
                    "type": "string",
                    "x-order": "4"
                },
                "descriptionUz": {
                    "type": "string",
                    "x-order": "5"
                },
                "descriptionRu": {
                    "type": "string",
                    "x-order": "6"
                },
                "descriptionEn": {
                    "type": "string",
                    "x-order": "7"
                },
                "position": {
                    "type": "integer",
                    "x-order": "8"
                },
                "groupId": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
...
        "models.ProductImages": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "1"
                },
                "productId": {
                    "type": "string",
                    "x-order": "2"
                },
                "image": {
                    "type": "string",
                    "x-order": "3"
                },
                "position": {
                    "type": "integer",
                    "x-order": "4"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "5"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "6"
                },
                "deletedAt": {
                    "type": "string",
                    "x-order": "7"
                }
            }
        },
...
        "models.ProductWithImages": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "1"
                },
                "groupId": {
                    "type": "string",
                    "x-order": "10"
                },
                "price": {
                    "type": "number",
                    "x-order": "11"
                },
                "seoDescription": {
                    "type": "string",
                    "x-order": "12"
                },
                "seoKeywords": {
                    "type": "string",
                    "x-order": "13"
                },
                "seoText": {
                    "type": "string",
                    "x-order": "14"
                },
                "seoTitle": {
                    "type": "string",
                    "x-order": "15"
                },
                "enabled": {
                    "type": "boolean",
                    "x-order": "16"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "17"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "18"
                },
                "deletedAt": {
                    "type": "string",
                    "x-order": "19"
                },
                "url": {
                    "type": "string",
                    "x-order": "2"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductImages"
                    },
                    "x-order": "21"
                },
                "nameUz": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameRu": {
                    "type": "string",
                    "x-order": "4"
                },
                "nameEn": {
                    "type": "string",
                    "x-order": "5"
                },
                "descriptionUz": {
                    "type": "string",
                    "x-order": "6"
                },
                "descriptionRu": {
                    "type": "string",
                    "x-order": "7"
                },
                "descriptionEn": {
                    "type": "string",
                    "x-order": "8"
                },
                "position": {
                    "type": "integer",
                    "x-order": "9"
                },
                "rate": {
                    "type": "integer"
                }
            }
        },
        "models.Products": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "1"
                },
                "groupId": {
                    "type": "string",
                    "x-order": "10"
                },
                "price": {
                    "type": "number",
                    "x-order": "11"
                },
                "seoDescription": {
                    "type": "string",
                    "x-order": "12"
                },
                "seoKeywords": {
                    "type": "string",
                    "x-order": "13"
                },
                "seoText": {
                    "type": "string",
                    "x-order": "14"
                },
                "seoTitle": {
                    "type": "string",
                    "x-order": "15"
                },
                "enabled": {
                    "type": "boolean",
                    "x-order": "16"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "17"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "18"
                },
                "deletedAt": {
                    "type": "string",
                    "x-order": "19"
                },
                "url": {
                    "type": "string",
                    "x-order": "2"
                },
                "nameUz": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameRu": {
                    "type": "string",
                    "x-order": "4"
                },
                "nameEn": {
                    "type": "string",
                    "x-order": "5"
                },
                "descriptionUz": {
                    "type": "string",
                    "x-order": "6"
                },
                "descriptionRu": {
                    "type": "string",
                    "x-order": "7"
                },
                "descriptionEn": {
                    "type": "string",
                    "x-order": "8"
                },
                "position": {
                    "type": "integer",
                    "x-order": "9"
                },
                "rate": {
                    "type": "integer"
                }
            }
        },
        "models.ProductsWithImagesAndPagination": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductWithImages"
                    },
                    "x-order": "1"
                },
                "page": {
                    "type": "integer",
                    "x-order": "2"
                },
                "pageSize": {
                    "type": "integer",
                    "x-order": "3"
                },
                "pageCount": {
                    "type": "integer",
                    "x-order": "4"
                },
                "count": {
                    "type": "integer",
                    "x-order": "5"
                }
            }
        },
...
        "models.UpdateProductInput": {
            "type": "object",
            "required": [
                "descriptionUz",
                "groupId",
                "nameUz",
                "price"
            ],
            "properties": {
                "groupId": {
                    "type": "string",
                    "x-order": "10"
                },
                "price": {
                    "type": "number",
                    "x-order": "11"
                },
                "seoDescription": {
                    "type": "string",
                    "x-order": "12"
                },
                "seoKeywords": {
                    "type": "string",
                    "x-order": "13"
                },
                "seoText": {
                    "type": "string",
                    "x-order": "14"
                },
                "seoTitle": {
                    "type": "string",
                    "x-order": "15"
                },
                "enabled": {
                    "type": "boolean",
                    "x-order": "16"
                },
                "url": {
                    "type": "string",
                    "x-order": "2"
                },
                "nameUz": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameRu": {
                    "type": "string",
                    "x-order": "4"
                },
                "nameEn": {
                    "type": "string",
                    "x-order": "5"
                },
                "descriptionUz": {
                    "type": "string",
                    "x-order": "6"
                },
                "descriptionRu": {
                    "type": "string",
                    "x-order": "7"
                },
                "descriptionEn": {
                    "type": "string",
                    "x-order": "8"
                },
                "position": {
                    "type": "integer",
                    "x-order": "9"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}